♻️ Refactor side navigation
In the a following change, we want to be able to have section titles in our side navigation.
Our current navigation was done with conditional logic recursing over the same component. This makes it difficult to know what level a heading is unless you have a counter counting the levels of recursion 🤔

Changes:
* Change heading colours to meet accessibility standards (make small text darker).
* Make level 2 headings a consistent font and size, regardless of whether it is a disclosure button.
* Correct semantic HTML for a menu. See reference: https://www.w3.org/WAI/ARIA/apg/patterns/menubar/examples/menubar-navigation/#sc_label
* Enable disclosures to open with enter key. See reference: https://headlessui.com/vue/disclosure#keyboard-interaction

Task: https://www.notion.so/centrapay/Sidebar-top-level-non-collapsible-sections-should-be-bigger-884638950dea4dd4940f7ae76addd238?pvs=4

Design review: https://centrapay.slack.com/archives/C04B93QUDSM/p1696825071503949

Test plan: Expect no breaking changes to existing navigation.
